gorm 创建表结构

+++ api/service.go
@@ -27,7 +27,7 @@ func NewService(db *databases.DB, conf *config.Config) *Service {
 type (
 	// ProxyRsp 代理信息结构体
 	ProxyRsp struct {
-		ID        int       `gorm:"column:id" json:"id"`
+		ID        uint64    `gorm:"column:id" json:"id"`
 		IP        string    `gorm:"column:ip" json:"ip"`
 		Port      int       `gorm:"column:port" json:"port"`
 		Schema    string    `gorm:"column:schema" json:"schema"`
